// desktop/src/ipc/generated/MachineSet.ts
// SPDX-License-Identifier: FSL-1.1-Apache-2.0
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SkuCheck } from "./SkuCheck";

/**
 * What `target machine` set.
 */
export type MachineSet = { name: string, sku: string, region: string | null, skuCheck: SkuCheck, };